requirements.txt
coinbaseadvanced==0.2.0
google-cloud-secret-manager==2.17.0
pyTelegramBotAPI==4.14.0
python-dotenv==0.17.1
python-json-logger==2.0.1
